requirements.txt
discord.py==1.7.3
python-dotenv==0.17.1
jishaku==2.0.0
aiosqlite==0.17.0
git+https://github.com/Rapptz/discord-ext-menus